reciever/requirements.txt
brother-ql-inventree==1.1
Flask==3.0.1
pdf2image==1.17.0
pyusb==1.2.1
requests==2.26.0
python-dotenv